backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain and LangGraph
langgraph==0.0.26
langchain==0.1.6
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langchain-community==0.0.20

# OpenAI
openai==1.12.0

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4
asyncpg==0.29.0

# Document processing
pypdf==4.0.1
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
